backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pypdf>=3.17.0
langchain>=0.0.350
langchain-google-genai>=0.0.6
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
python-multipart>=0.0.6
google-generativeai>=0.3.0
pydantic>=2.0.0

# For hierarchical extraction
langextract>=0.1.0
